# backend/docs/swagger.yaml
definitions:
...
  github_com_raids-lab_crater_dao_model.BuildSource:
    enum:
    - Dockerfile
    - PipApt
    - Snapshot
    - EnvdAdvanced
    - EnvdRaw
    type: string
    x-enum-varnames:
    - Dockerfile
    - PipApt
    - Snapshot
    - EnvdAdvanced
    - EnvdRaw
...
  github_com_raids-lab_crater_internal_util.VolumeMount:
    properties:
      datasetID:
        type: integer
      mountPath:
        type: string
      subPath:
        type: string
      type:
        $ref: '#/definitions/github_com_raids-lab_crater_internal_util.VolumeType'
    type: object
  github_com_raids-lab_crater_internal_util.VolumeType:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - _
    - FileType
    - DataType
...
  internal_handler_image.CreateByDockerfileRequest:
    properties:
      archs:
        items:
          type: string
        type: array
      description:
        type: string
      dockerfile:
        type: string
      name:
        type: string
      tag:
        type: string
      tags:
        items:
          type: string
        type: array
      template:
        type: string
      volumeMounts:
        items:
          $ref: '#/definitions/github_com_raids-lab_crater_internal_util.VolumeMount'
        type: array
    type: object
  internal_handler_image.CreateByEnvdRequest:
    properties:
      archs:
        items:
          type: string
        type: array
      base:
        type: string
      buildSource:
        $ref: '#/definitions/github_com_raids-lab_crater_dao_model.BuildSource'
      description:
        type: string
      envd:
        type: string
      name:
        type: string
      python:
        type: string
      tag:
        type: string
      tags:
        items:
          type: string
        type: array
      template:
        type: string
    type: object
  internal_handler_image.CreateKanikoRequest:
    properties:
      archs:
        items:
          type: string
        type: array
      description:
        type: string
      image:
        type: string
      name:
        type: string
      packages:
        type: string
      requirements:
        type: string
      tag:
        type: string
      tags:
        items:
          type: string
        type: array
      template:
        type: string
    type: object
...
info:
  contact: {}
  description: This is the API server for Crater, a Multi-tenant AI Model Training
    Platform based on Kubernetes.
  title: Crater API
  version: 1.0.0
paths:
...
  /v1/admin/images/kaniko:
    get:
      consumes:
      - application/json
      description: 管理员获取所有镜像制作信息
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: 管理员获取相关镜像的功能
      tags:
      - ImagePack
    post:
      consumes:
      - application/json
      description: 获取参数，生成变量，调用接口
      parameters:
      - description: 创建ImagePack
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_handler_image.CreateKanikoRequest'
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: 创建ImagePack CRD和数据库kaniko entity
      tags:
      - ImagePack
...
  /v1/images/dockerfile:
    post:
      consumes:
      - application/json
      description: 获取参数，提取Dockerfile中的基础镜像，调用接口
      parameters:
      - description: 创建ImagePack CRD
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_handler_image.CreateByDockerfileRequest'
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: 接受用户传入的Dockerfile和描述，创建镜像
      tags:
      - ImagePack
  /v1/images/envd:
    post:
      consumes:
      - application/json
      description: 获取参数，提取Dockerfile中的基础镜像，调用接口
      parameters:
      - description: 创建ImagePack CRD
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_handler_image.CreateByEnvdRequest'
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: 接受用户传入的Envd内容和描述，创建镜像
      tags:
      - ImagePack
...
  /v1/images/kaniko:
    get:
      consumes:
      - application/json
      description: 返回该用户所有的镜像构建数据
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: 用户获取镜像构建信息
      tags:
      - ImagePack
    post:
      consumes:
      - application/json
      description: 获取参数，生成变量，调用接口
      parameters:
      - description: 创建ImagePack CRD & Kaniko entity
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_handler_image.CreateKanikoRequest'
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: 创建ImagePack CRD和数据库Kaniko entity
      tags:
      - ImagePack
...
securityDefinitions:
  Bearer:
    description: 访问 /login 并获取 TOKEN 后，填入 'Bearer ${TOKEN}' 以访问受保护的接口
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"